Fix a "non-existing property" QML error, makes plasma output less noisy
The "action" property was commented out, making the
"Component.onDestruction" throw a lot of errors about "action" property
not existing.

The shortcutActions system doesn't seem to be existing yet, so for now
the property is back but it's always set to "null".

Given the property can be "null" according to the commented code, we
should check for that in "Component.onDestruction" anyway. This makes
the error outputs go away.