Make Oxygen the default font
The startkde portion contains the bits to write out a kdeglobals default
file if it doesn't exist with font settings applied. Usually, with
install  prefix set to /usr, the installed oxygen font is found automatically
by  fontconfig. If we're installed to a different prefix, we need to point
fontconfig at the font. We do that by linking it from either
XDG_DATA_HOME or ~/.fonts/ and updating fontconfig with it. The latter
is irrelevant for systems that install into /usr.

REVIEW:116625