Remove the notification timers from DataEngine
The popups have their own timers to hide themselves and these are paused
when user hovers over the popup; syncing the popup and DataEngine's
timers would be unnecessary difficult so we leave only the popup one.

This also makes sure that persistent notifications' actions will keep
working as long as the source KNotification object still exists.