Remove the notifications plasmoid tabbar
Now only persistent notifications will be shown in the history and they
will be merged into one if it comes from the same app and same title.
This should significantly reduce the amount of notifications in the
popup -> no tabs needed.

Alternatively a StatusNotifierItem should be created if the developer
wishes to give the user an indication that something important happened
and needs action.