Fix search input interaction
This reworks the focus handling of the search input.

Previously, the lineedit would never keep focus, the focus always stayed
with the root item. This means that the keys for the lineedit needed to
be filtered one by one, and the lineedit wouldn't get the proper focus
visually.

This patch changes the mechanism to switch focus between lineedit and
root view / tab view. It makes the lineedit behave as one is used to.

As a cherry on top: Kickoff will paste the clipboard content with
Control+V even if the search input doesn't have focus.

BUG:331960