Port away from MouseEventsListener::onClicked.
MouseEventListener's state tracking for a press+release pair seems
to get confused when the mouse grab changes mid-pair due to the
context menu opening. This is adds to the list of unresolved issues
in MouseEventListener. Since we're already handling presses and
releases ourselves for other reasons we don't really need the helper.

BUG:332502