Discard all deleted before constructing the Scene
The Scene connects to the deletedRemoved signal and tries to remove the
Deleted from it's list of Windows, which will fail because a SceneWindow
doesn't get created for an existing Deleted in the Scene setup.

It doesn't make sense to create a SceneWindow for a Deleted as there is
no pixmap around anyway. Thus the best approach is to ensure that there
are no Deleted by just discarding all prior to creating the Scene.

BUG: 333275
CCBUG: 334273
REVIEW: 117557

"cherry-picked" from kwin.git 8e631028552c4b474a317ebf589e31c13289c6fd
Author: mgraesslin
Date: Mon, 14 Apr 2014 09:19:43 +0200